* theme.c (Theme_Build_Objects): Use style 'gpe_labels', not 'gpe_ownerinfo_labels'
* ipaqscreen/main.c (ipaqscreen_Build_Objects): Make the GUI a bit
nicer.  Wrap a scrolledwindow/viewport around table and return
that.  Fix padding and justification. Use gpe_labels style for the
labels in the left table column.

* theme.c (Theme_Build_Objects): Make the GUI a bit nicer.  Wrap a
scrolledwindow/viewport around table and return that.  Fix padding
and justification. Use gpe_ownerinfo_labels style for the labels
in the left table column.