Add support for h2200.
Also, turn echo off; if the bluetooth chip emits characters on power up,
they would be echoed back to the chip, confusing it. This should only
be necessary to do once, as hciattach turns echo off too (but too
late the first time, since the bluetooth chip is powered up when the
port is opened, before hciattach gets a chance to turn echo off).